moves and renames generic command handler
