+++ src/applications/tribute-tools/handlers/index.ts
@@ -1,1 +1,2 @@
+export * from './interactionExecuteHandler';
 export * from './sweep';
