+++ src/applications/tribute-tools/handlers/sweep/index.ts
@@ -1,3 +1,2 @@
 export * from './sweepEndedPollsHandler';
-export * from './sweepInteractionExecuteHandler';
 export * from './sweepPollReactionHandler';
